pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "django-routers>=0.2",
    "docker>=7.1.0",
    "fastapi>=0.115.12",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "gitpython>=3.1.44",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "oauthlib>=3.2.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
